Migrate fully to enums
This commit removes many of the string constants that described types of actors that could be spawned. Actors that Fishy has not yet implemented can still be spawned via the `SpawnActor(string, vec3, vec3)` overloaded method of `Spawner`. A position is required because Fishy will not be able to replicate the vanilla spawning algorithm for an actor type it doesn't know about yet.